Added extensions for VC w/ collection views
