+++ Clicker/View Controllers/CardController.swift
@@ -77,7 +77,6 @@ class CardController: UIViewController, UICollectionViewDelegate, UICollectionVi
     }
    
     // MARK - NAME THE POLL
-    
     func setupName() {
         nameView = NameView()
         nameView.sessionId = sessionId
@@ -249,64 +248,47 @@ class CardController: UIViewController, UICollectionViewDelegate, UICollectionVi
         }
     }
     
-    // MARK: - COLLECTIONVIEW
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        if (collectionView == verticalCollectionView) {
-            return datePollsArr.count
-        } else { // mainCollectionView
-            return currentPolls.count
+    // MARK: - Vertical Collection View
+    func setupVertical() {
+        mainCollectionView.removeFromSuperview()
+        zoomOutButton.removeFromSuperview()
+        countLabel.removeFromSuperview()
+        
+        setupVerticalNavBar()
+        setupVerticalCollectionView()
+    }
+    
+    func setupVerticalCollectionView() {
+        let layout = UICollectionViewFlowLayout()
+        verticalCollectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
+        layout.minimumLineSpacing = 10
+        layout.scrollDirection = .vertical
+        verticalCollectionView.delegate = self
+        verticalCollectionView.dataSource = self
+        verticalCollectionView.register(CardDateCell.self, forCellWithReuseIdentifier: dateIdentifier)
+        verticalCollectionView.showsVerticalScrollIndicator = false
+        verticalCollectionView.showsHorizontalScrollIndicator = false
+        verticalCollectionView.alwaysBounceVertical = true
+        verticalCollectionView.backgroundColor = .clear
+        verticalCollectionView.isPagingEnabled = true
+        view.addSubview(verticalCollectionView)
+        view.sendSubview(toBack: verticalCollectionView)
+        
+        verticalCollectionView.snp.makeConstraints { make in
+            make.edges.equalToSuperview()
         }
     }
     
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        if (collectionView == verticalCollectionView) {
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: dateIdentifier, for: indexPath) as! CardDateCell
-            let poll = datePollsArr[indexPath.item].1.first
-            cell.poll = poll
-            cell.date = datePollsArr[indexPath.item].0
-            cell.userRole = userRole
-            cell.cardType = getCardType(from: poll!)
-            cell.configure()
-            return cell
-        }
-         // mainCollectionView
-        switch (userRole) {
-        case .admin:
-            let poll = currentPolls[indexPath.item]
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: askedIdentifer, for: indexPath) as! AskedCard
-            cell.socket = socket
-            socket.addDelegate(cell)
-            cell.poll = poll
-            cell.endPollDelegate = self
-            cell.cardType = getCardType(from: poll)
-            cell.configure()
-            return cell
-        default:
-            let poll = currentPolls[indexPath.item]
-            let cell = collectionView.dequeueReusableCell(withReuseIdentifier: answerIdentifier, for: indexPath) as! AnswerCard
-            cell.socket = socket
-            socket.addDelegate(cell)
-            cell.poll = poll
-            cell.cardType = getCardType(from: poll)
-            cell.configure()
-            return cell
-        }
+    func setupVerticalNavBar() {
+        self.navigationController?.navigationBar.isTranslucent = true
+        self.navigationItem.titleView = UIView()
+        self.navigationItem.rightBarButtonItems = []
     }
     
-    func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
-        if (collectionView == mainCollectionView) {
-            // UPDATE COUNT LABEL
-            let countString = "\(indexPath.item + 1)/\(currentPolls.count)"
-            countLabel.attributedText = getCountLabelAttributedString(countString)
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        if (collectionView == mainCollectionView) {
-            return CGSize(width: view.frame.width * 0.9, height: mainCollectionView.frame.height)
-        } else {
-            return CGSize(width: view.frame.width * 0.76, height: 440)
-        }
+    func revertToHorizontal() {
+        verticalCollectionView.removeFromSuperview()
+        setupCards()
+        setupHorizontalNavBar()
     }
 
     // MARK: Get CardType
