+++ Clicker/View Controllers/PollsViewController+Extension.swift
@@ -1,0 +1,38 @@
+//
+//  PollsViewController+Extension.swift
+//  Clicker
+//
+//  Created by Kevin Chan on 8/26/18.
+//  Copyright © 2018 CornellAppDev. All rights reserved.
+//
+
+import UIKit
+
+extension PollsViewController {
+    
+    // MARK: - COLLECTIONVIEW
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 2
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: pollsIdentifier, for: indexPath) as! PollsCell
+        cell.editSessionDelegate = self
+        cell.parentNavController = self.navigationController
+        if (indexPath.item == 0) {
+            cell.pollType = .created
+        } else {
+            cell.pollType = .joined
+        }
+        return cell
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: pollsCollectionView.frame.width, height: pollsCollectionView.frame.height)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
+        pollsOptionsView.collectionView.selectItem(at: indexPath, animated: false, scrollPosition: .centeredVertically)
+    }
+    
+}
