+++ Clicker/View Controllers/PollsViewController.swift
@@ -30,32 +30,6 @@ class PollsViewController: UIViewController, UICollectionViewDelegate, UICollect
         setupConstraints()
     }
     
-    // MARK: - COLLECTIONVIEW
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 2
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: pollsIdentifier, for: indexPath) as! PollsCell
-        cell.editSessionDelegate = self
-        cell.parentNavController = self.navigationController
-        if (indexPath.item == 0) {
-            cell.pollType = .created
-        } else {
-            cell.pollType = .joined
-        }
-        print("row: \(indexPath.row)\nitem: \(indexPath.item) \nsection: \(indexPath.section)")
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: pollsCollectionView.frame.width, height: pollsCollectionView.frame.height)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
-        pollsOptionsView.collectionView.selectItem(at: indexPath, animated: false, scrollPosition: .centeredVertically)
-    }
-    
     // MARK: - SCROLLVIEW
     func scrollViewDidScroll(_ scrollView: UIScrollView) {
         pollsOptionsView.sliderBarLeftConstraint.constant = scrollView.contentOffset.x / 2
